third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/serviceworker/fetch-request-fallback-expected.txt

CONSOLE ERROR: XMLHttpRequest cannot load http://localhost:8000/serviceworker/resources/fetch-access-control.php?. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: XMLHttpRequest cannot load http://localhost:8000/serviceworker/resources/fetch-access-control.php?. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: Access to Image at 'http://localhost:8000/serviceworker/resources/fetch-access-control.php?PNGIMAGE&' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: Access to Image at 'http://localhost:8000/serviceworker/resources/fetch-access-control.php?PNGIMAGE&' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
This is a testharness.js-based test.
PASS Verify the fallback behavior of FetchEvent 
Harness: the test ran to completion.
